Feat : 게시글 확대 기능, 확대 화면 추가 및 post 개발 수정, postview 화면 추가 및 화면 수정

+++ src/main/java/com/hobbyproject/service/PostServiceImpl.java
@@ -2,7 +2,9 @@ package com.hobbyproject.service;
 
 import com.hobbyproject.dto.post.request.PostEditDto;
 import com.hobbyproject.dto.post.request.PostWriteDto;
+import com.hobbyproject.entity.Member;
 import com.hobbyproject.entity.Post;
+import com.hobbyproject.repository.MemberRepository;
 import com.hobbyproject.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -15,13 +17,18 @@ import java.util.List;
 public class PostServiceImpl implements PostService {
 
     private final PostRepository postRepository;
+    private final MemberRepository memberRepository;
 
     @Override
     @Transactional
-    public void postCreate(PostWriteDto postWriteDto) {
+    public void postCreate(PostWriteDto postWriteDto,Long id) {
+
+        Member member = memberRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+
         Post post=Post.builder()
                 .title(postWriteDto.getTitle())
                 .content(postWriteDto.getContent())
+                .member(member)
                 .build();
 
         postRepository.save(post);
@@ -29,8 +36,15 @@ public class PostServiceImpl implements PostService {
 
     @Override
     @Transactional
-    public void postEdit(PostEditDto postEditDto) {
+    public void postEdit(PostEditDto postEditDto,Long id) {
+
+        Member member = memberRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+
         Post post = postRepository.findById(postEditDto.getPostId()).orElseThrow(IllegalArgumentException::new);
+
+        if (!post.getMember().equals(member)) {
+            throw new IllegalArgumentException("회원이 일치하지 않습니다.");
+        }
 
         post.edit(postEditDto);
 
@@ -39,8 +53,15 @@ public class PostServiceImpl implements PostService {
 
     @Override
     @Transactional
-    public void postDelete(Long postId) {
+    public void postDelete(Long postId,Long id) {
+
+        Member member = memberRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+
         Post post = postRepository.findById(postId).orElseThrow(IllegalArgumentException::new);
+
+        if (!post.getMember().equals(member)) {
+            throw new IllegalArgumentException("회원이 일치하지 않습니다.");
+        }
 
         postRepository.delete(post);
     }
@@ -48,5 +69,21 @@ public class PostServiceImpl implements PostService {
     @Override
     public List<Post> findPosts() {
         return postRepository.findAll();
+    }
+
+    @Override
+    public Post findPost(Long postId) {
+        return postRepository.findById(postId).orElseThrow(IllegalArgumentException::new);
+    }
+
+    @Override
+    public boolean postMemberCheck(Post post, Long id) {
+
+        Member member = memberRepository.findById(id).orElseThrow(IllegalArgumentException::new);
+
+        if(post.getMember().equals(member)){
+            return true;
+        }
+        return false;
     }
 }
